+++ src/fintoolsom/__init__.py
@@ -1,5 +1,6 @@
 from .__version__ import __version__
-from .fixedIncome import CLBond, Bond, Coupons, Coupon, NelsonSiegelSvensson
+from .fixedIncome import CLBond, Bond, Coupons, Coupon
 from .fixedIncome import get_irr
+from .models import NelsonSiegelSvensson
 from .rates import Rate, RateConvention, CompoundedInterestConvention, LinearInterestConvention, ExponentialInterestConvention
 from .dates import ActualDayCountConvention, Days30ADayCountConvention, Days30EDayCountConvention, Days30EISDADayCountConvention, Days30UDayCountConvention
