+++ src/fintoolsom/fixedIncome/__init__.py
@@ -1,5 +1,4 @@
 from .Bonds import Bond, Coupons, Coupon
 from .CLBonds import CLBond
 from .Deposits import Deposit
-from .NelsonSiegelSvensson import NelsonSiegelSvensson
 from .utils import get_irr
