+++ src/fintoolsom/models/__init__.py
@@ -1,0 +1,1 @@
+from .NelsonSiegelSvensson import NelsonSiegelSvensson
